Moving cell-mouse model to correct folder
